Create Enhanced_K-means_algorithm.py
Cluster analysis is one of the primary data 
analysis methods and k-means is one of the most well 
known popular clustering algorithms. The k-means 
algorithm is one of the frequently used clustering method 
in data mining, due to its performance in clustering 
massive data sets. The final clustering result of the k
means clustering algorithm greatly depends upon the 
correctness of the initial centroids, which are selected 
randomly. The original k-means algorithm converges to 
local minimum, not the global optimum. Many 
improvements were already proposed to improve the 
performance of the k-means, but most of these require 
additional inputs like threshold values for the number of 
data points in a set. In this paper a new method is 
proposed for finding the better initial centroids and to 
provide an efficient way of assigning the data points to 
suitable clusters with reduced time complexity. According 
to our experimental results, the proposed algorithm has 
the more accuracy with less computational time 
comparatively original k-means clustering algorithm.

https://citeseerx.ist.psu.edu/document?repid=rep1&type=pdf&doi=5b4d09f41f8fa28c8f4ac3e7b8a474ae9f84b197